Variable Explorer: Rename widgets/varexp to widgets/variableexplorer
