+++ spyderlib/plugins/variableexplorer.py
@@ -7,7 +7,7 @@
 """Variable Explorer Plugin"""
 
 from spyderlib.qt.QtGui import QGroupBox, QStackedWidget, QVBoxLayout, QWidget
-from spyderlib.qt.QtCore import Signal, Slot
+from spyderlib.qt.QtCore import Signal
 import spyderlib.utils.icon_manager as ima
 
 # Local imports
@@ -16,7 +16,7 @@ from spyderlib.config.main import CONF
 from spyderlib.utils import programs
 from spyderlib.plugins import SpyderPluginMixin, PluginConfigPage
 from spyderlib.widgets.externalshell.monitor import REMOTE_SETTINGS
-from spyderlib.widgets.varexp.namespacebrowser import NamespaceBrowser
+from spyderlib.widgets.variableexplorer.namespacebrowser import NamespaceBrowser
 
 
 class VariableExplorerConfigPage(PluginConfigPage):
