+++ spyderlib/widgets/variableexplorer/arrayeditor.py
