+++ spyderlib/widgets/variableexplorer/collectionseditor.py
@@ -37,17 +37,17 @@ from spyderlib.config.base import _
 from spyderlib.config.gui import get_font
 from spyderlib.utils.misc import fix_reference_name
 from spyderlib.utils.qthelpers import add_actions, create_action, qapplication
-from spyderlib.widgets.varexp.utils import (sort_against, get_size,
+from spyderlib.widgets.variableexplorer.utils import (sort_against, get_size,
                get_human_readable_type, value_to_display, get_color_name,
                is_known_type, FakeObject, Image, ndarray, array, MaskedArray,
                unsorted_unique, try_to_eval, is_editable_type, DataFrame,
                Series, display_to_value, np_savetxt)
 if ndarray is not FakeObject:
-    from spyderlib.widgets.varexp.arrayeditor import ArrayEditor
+    from spyderlib.widgets.variableexplorer.arrayeditor import ArrayEditor
 if DataFrame is not FakeObject:
-    from spyderlib.widgets.varexp.dataframeeditor import DataFrameEditor
-from spyderlib.widgets.varexp.texteditor import TextEditor
-from spyderlib.widgets.varexp.importwizard import ImportWizard
+    from spyderlib.widgets.variableexplorer.dataframeeditor import DataFrameEditor
+from spyderlib.widgets.variableexplorer.texteditor import TextEditor
+from spyderlib.widgets.variableexplorer.importwizard import ImportWizard
 from spyderlib.py3compat import (to_text_string, is_text_string, PY3, io,
                                  is_binary_string, getcwd)
 
@@ -1165,7 +1165,7 @@ class CollectionsEditorTableView(BaseTableView):
     def oedit(self, key):
         """Edit item"""
         data = self.model.get_data()
-        from spyderlib.widgets.varexp.objecteditor import oedit
+        from spyderlib.widgets.variableexplorer.objecteditor import oedit
         oedit(data[key])
 
     def plot(self, key, funcname):
