+++ spyderlib/widgets/variableexplorer/texteditor.py
