+++ spyderlib/widgets/variableexplorer/utils.py
